+++ contracts/strategy/arbitrary/Arbitrary.sol
@@ -12,6 +12,7 @@ contract Arbitrary is StrategyCommon {
      @param oneToken_ bind this instance to one oneToken vault
      @param description_ metadata has no impact on logic
      */
+
     constructor(address oneTokenFactory_, address oneToken_, string memory description_) 
         StrategyCommon(oneTokenFactory_, oneToken_, description_)
     {}
@@ -24,7 +25,7 @@ contract Arbitrary is StrategyCommon {
     @param signature the function signature
     @param data abi-encodeded bytecode of the parameter values to send
     */
-    function executeTransaction(address target, uint value, string memory signature, bytes memory data) external onlyOwner returns (bytes memory) {
+    function executeTransaction(address target, uint256 value, string memory signature, bytes memory data) external onlyOwner returns (bytes memory) {
         bytes memory callData;
 
         if (bytes(signature).length == 0) {
